Updated in club route

+++ src/controllers/clubController.ts
@@ -81,11 +81,22 @@ clubController.post("/addAthlete", userValidation, async (req: RequestWithUser, 
 });
 
 /********************************
-    ADD Chairperson
+    ADD Chairperson 
  *******************************/
 clubController.post("/addChair", userValidation, async (req: RequestWithUser, res) => {
   try {
     const info = req.body.info;
+    const user = req.user!;
+
+    //checks role permissions. if not the proper role, throws error.
+    if (info.role === "chair") {
+      await roleValidator(user.id!, info.club_id, ["chair"], "clubs_users");
+    } else if (info.role === "vice_chair") {
+      await roleValidator(user.id!, info.club_id, ["chair", "vice_chair"], "clubs_users");
+    } else {
+      throw new CustomError(401, "That role cannot be added through this route.");
+    }
+
     //finds user by email.
     const clubMemberResults = await pool.query("SELECT * FROM users WHERE email = $1", [info.email]);
     if (clubMemberResults.rowCount === 0) {
@@ -120,7 +131,7 @@ clubController.get("/getClubMembers/:id", userValidation, async (req: RequestWit
   try {
     const club_id = +req.params.id;
     const user = req.user!;
-    //checks that updater is chair, ice_chair, or athlete for club, if not, throws error.
+    //checks that updater is chair, vice_chair, or athlete for club, if not, throws error.
     await roleValidator(user.id!, club_id, ["chair", "vice_chair", "athlete"], "clubs_users");
 
     const athletesResults = await pool.query(
@@ -209,10 +220,11 @@ clubController.put("/updateChairperson", userValidation, async (req: RequestWith
     //checks that updater is chair for club, if not, throws error.
     await roleValidator(user.id!, info.club_id, ["chair"], "clubs_users");
 
-    //Utility function to get query arguments
-    const [queryString, valArray] = getQueryArgs("update", "clubs_users", info, info.club_id);
     //Pass UPDATE to DB
-    const result = await pool.query(queryString, valArray);
+    const result = await pool.query(
+      `UPDATE clubs_users SET role = $1 WHERE club_id = $2 AND user_id = $3 RETURNING *`,
+      [info.role, info.club_id, info.user_id]
+    );
     const updatedClubsUsersItem = result.rows[0];
 
     res.status(200).json({ message: "Role Updated", updatedClubsUsersItem });
@@ -260,9 +272,10 @@ clubController.put("/updateClub", userValidation, async (req: RequestWithUser, r
 **************************/
 clubController.delete("/removeSelf/:id", userValidation, async (req: RequestWithUser, res) => {
   try {
-    const club_id = req.params;
+    const club_id = req.params.id;
     const user = req.user!;
-    //checks that updater is chair or manager on team, if not, throws error.
+
+    //checks that user is on team/club, if not, throws error.
     await roleValidator(user.id!, +club_id!, ["chair", "vice_chair", "athlete"], "clubs_users");
 
     //Ensures that at least one chair remains on the club.
@@ -282,7 +295,7 @@ clubController.delete("/removeSelf/:id", userValidation, async (req: RequestWith
 
     const removed = results.rows[0];
 
-    res.status(200).json({ message: "`Removed from club.", removed });
+    res.status(200).json({ message: "Removed from club.", removed });
   } catch (error) {
     console.log(error);
     if (error.status < 500) {
@@ -301,22 +314,25 @@ clubController.delete("/removeAthlete", userValidation, async (req: RequestWithU
     const info = req.query;
     const user = req.user!;
     //checks that updater is chair or manager on team, if not, throws error.
-    await roleValidator(user.id!, +info.club_id!, ["chair", "vice_chair"], "clubs_users");
+    await roleValidator(+user.id!, +info.club_id!, ["chair", "vice_chair"], "clubs_users");
 
     //Send DELETE query to DB
-    const results = await pool.query(
-      "DELETE FROM clubs_users WHERE club_id = $1 AND user_id = $2 AND role = $3",
-      [info.club_id, user.id, "athlete"]
-    );
+    const [queryString, valArray] = getQueryArgs("delete", "clubs_users", info);
 
-    const removed = results.rows[0];
+    const results = await pool.query(queryString, valArray);
 
-    res.status(200).json({ message: "Removed from club.", removed });
+    if (results.rowCount === 0) {
+      throw new CustomError(404, "Clubmember not deleted. Could not find clubmember.");
+    }
+
+    res.status(200).json({ message: "Removed from club." });
   } catch (error) {
-    res.status(500).json({
-      message: "Error. Athlete failed to be removed.",
-      error,
-    });
+    console.log(error);
+    if (error.status < 500) {
+      res.status(error.status).json({ message: error.message });
+    } else {
+      res.status(500).json({ message: "Internal server error", error });
+    }
   }
 });
 
@@ -330,6 +346,15 @@ clubController.delete("/removeChairperson", userValidation, async (req: RequestW
     //checks that deleter is chair, if not, throws error.
     await roleValidator(user.id!, +info.club_id!, ["chair"], "clubs_users");
 
+    //Ensures that at least one chair remains on the club.
+    const chairResults = await pool.query("SELECT * FROM clubs_users WHERE club_id = $1 AND role = $2;", [
+      +info.club_id!,
+      "chair",
+    ]);
+    if (chairResults.rowCount === 1 && chairResults.rows[0].user_id === user.id) {
+      throw new CustomError(403, "Must be atleast one chair on club.");
+    }
+
     //Send DELETE query to DB
     const results = await pool.query("DELETE FROM clubs_users WHERE club_id = $1 AND user_id = $2;", [
       info.club_id,
@@ -340,10 +365,12 @@ clubController.delete("/removeChairperson", userValidation, async (req: RequestW
 
     res.status(200).json({ message: "Removed from club.", removed });
   } catch (error) {
-    res.status(500).json({
-      message: "Error. Member failed to be removed.",
-      error,
-    });
+    console.log(error);
+    if (error.status < 500) {
+      res.status(error.status).json({ message: error.message });
+    } else {
+      res.status(500).json({ message: "Internal server error", error });
+    }
   }
 });
 
@@ -355,19 +382,21 @@ clubController.delete("/removeClub", userValidation, async (req: RequestWithUser
     const info = req.query;
     const user = req.user!;
 
-    await roleValidator(user.id!, +info.club_id!, ["chair"], "clubs_users");
+    await roleValidator(+user.id!, +info.id!, ["chair"], "clubs_users");
 
     //Send DELETE query to users table in DB
-    const results = await pool.query("DELETE FROM clubs WHERE id = $1", [info.club_id]);
+    const results = await pool.query("DELETE FROM clubs WHERE id = $1", [info.id]);
 
     const removed = results.rows[0];
 
     res.status(200).json({ message: "Club Removed", removed });
   } catch (error) {
-    res.status(500).json({
-      message: "Error. Club failed to delete.",
-      error,
-    });
+    console.log(error);
+    if (error.status < 500) {
+      res.status(error.status).json({ message: error.message });
+    } else {
+      res.status(500).json({ message: "Internal server error", error });
+    }
   }
 });
